Made change to convert all internal Integers to Double. This is more compliant with the XPath spec and will solve the > / < problems of ClassCastExceptions.

+++ src/java/main/org/jaxen/function/LastFunction.java
@@ -25,8 +25,8 @@ public class LastFunction implements Function
         throw new FunctionCallException( "last() requires no arguments." );
     }
 
-    public static Integer evaluate(Context context)
+    public static Double evaluate(Context context)
     {
-        return new Integer( context.getSize() );
+        return new Double( context.getSize() );
     }
 }
